rename Value prop to Blurb
